Move AbstractOfflineAction to script implementations package
