+++ src/games/stendhal/server/script/OfflineClearSentence.java
@@ -1,10 +1,19 @@
-/**
- *
- */
+/***************************************************************************
+ *                 Copyright © 2014-2024 - Faiumoni e. V.                  *
+ ***************************************************************************
+ ***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ***************************************************************************/
 package games.stendhal.server.script;
 
 import java.util.List;
 
+import games.stendhal.server.core.scripting.impl.AbstractOfflineAction;
 import games.stendhal.server.entity.player.Player;
 import marauroa.common.game.RPObject;
